+++ packages/amplify-utils/src/index.js
@@ -1,0 +1,9 @@
+/* istanbul ignore if */
+if (!Error.prepareStackTrace) {
+	require('source-map-support/register');
+}
+
+export * from './fs';
+export * from './path';
+export * from './util';
+